bug: 랭킹표시 오류 확인
